feat: se agregan nuevos endpoints a la api usuario, con paginación fix: se corrije el tipo de dato en Rol

+++ usuario/src/main/java/cl/duoc/lunari/api/user/repository/RoleRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import cl.duoc.lunari.api.user.model.UserRole;
 
 @Repository
-public interface RoleRepository extends JpaRepository<UserRole, Long> {
+public interface RoleRepository extends JpaRepository<UserRole, Integer> {
 	
 }
